chore: remove cli as it is not used anywhere
