--- test/sharness/.gitignore
@@ -1,5 +1,0 @@
-!/lib/
-bin/
-lib/sharness/
-test-results/
-trash directory.*.sh/
